1821 - Contains a role to build a self-extracting archive containing files to install Winlogbeat, Sysmon, and GRR on Windows. Adds a kibana dashboard for Winlogbeat.
Exposed logstash service.
Changed winlogbeat to send beats to logstash rather than elasticsearch.

In anticipation of specifying a specifying a PFSense Firewall IP address and port, the
Logstash URL and Port are now specified on the command line when building the installer.

Modified to handle requests from front-end.

Modified to handle the new way the api sends data.