Added cookie storing for build type (MDT/CPT)
The web interface will store the last known build type (CPT or MDT) and default back to that build type unless otherwise selected.  If no previous type was stored it will default to CPT
Added putty package

Sneaking in putty package addition to this merge as it was added to 3.5 and was supposed to be added to devel but no one ever approved it.