THISISCVAH-11670 - Suricata install defects
Disconnected interfaces are disabled when installing zeek, suricata or arkime.  This was added because suricata fails to install the suricata template on helm install with a disconnected interface.

Added fixes to Robot Framework so it selects the first non disconnected interface.
Added fixes to hardware pipeline for zeek rules and interface naming.

Co-authored-by: Brandon Clarke <brandon.clarke@amstechnologiescorp.com>