THISISCVAH-12005 Update Manifest for App store archival items
Updated manifest to pull archived items directly from the AppStore folder to prevent future duplication when archiving items. Moved CPT items to the archived list according to the approved archived list.  Created an archive list for the MDT section and added RHEL Server 7.8 to the archive list for MDT. Added the "archive" term to the split path on the manifest.py file so the archive section is treated the same as the appstore section