Update ruleset to remove assigned sensors
update the rulesets to remove assigned sensors that have had the app (Suricata or Zeek) uninstalled and update the status of the sync'd sensors.

Sync would fail if user left the catalog page before uninstall of app was completed so moved it to backend.

updated unit testing and fixed bug that was not removing nodes after a sensor uninstall from k8s that were stuck in a pending cycle